build: configure Gradle build system, fix C++ includes/flags, and verify 100% test pass in WSL2
- Add root build.gradle, settings.gradle, gradle.properties, and app/build.gradle with Android Gradle Plugin 8.3.2 and Gradle wrapper 8.7.
- Fix CMake target include directories and set CMAKE_POSITION_INDEPENDENT_CODE for shared JNI library.
- Fix missing cstring headers, signed/unsigned warnings, and AmlogicAdapter simulation fallback.
- Add reflection-based TvInputHardwareInfo detection for standard Android SDK compatibility.
- 90/90 unit tests passing (100%) and successful APK assembly (debug and release).

+++ bridge/Cs378xClient.cpp
@@ -684,7 +684,7 @@ bool Cs378xClient::readEncryptedPacket(int socketFd, uint8_t& outCmd, uint16_t& 
     }
 
     outPayload.clear();
-    if (payloadLen > 0 && 20 + payloadLen <= totalPadded) {
+    if (payloadLen > 0 && static_cast<size_t>(20 + payloadLen) <= totalPadded) {
         outPayload.assign(&allDec[20], &allDec[20 + payloadLen]);
     }
     return true;
